[level 1] Title: 부족한 금액 계산하기, Time: 0.07 ms, Memory: 62.1 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.kt
@@ -5,7 +5,7 @@ class Solution {
         for (i in 1..count) {
             total += (price * i)
         }
-        answer = -((money - total).toLong())
+        answer = -(money - total)
         if (answer <= 0) return 0
         else return answer
     }
